chore(refactor): migrate ThemeJson and SectionNames to Domain namespace, remove TJGConfig, and update references across the codebase
