+++ docs/02-advanced-usage.md
@@ -43,8 +43,8 @@ declare(strict_types=1);
 
 namespace YourVendor\YourProject;
 
-use ItalyStrap\ThemeJsonGenerator\Application\Config\ThemeJson;
 use ItalyStrap\ThemeJsonGenerator\Domain\Input\Settings\Presets;
+use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\ThemeJson;
 
 return static function (ThemeJson $themeJson, Presets $presets): void {
     // ...
@@ -297,9 +297,7 @@ declare(strict_types=1);
 
 namespace YourVendor\YourProject;
 
-use ItalyStrap\ThemeJsonGenerator\Application\Config\ThemeJson;
-use ItalyStrap\ThemeJsonGenerator\Domain\Input\Settings\Presets;
-use Psr\Container\ContainerInterface;
+use ItalyStrap\ThemeJsonGenerator\Domain\Input\Settings\Presets;use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\ThemeJson;use Psr\Container\ContainerInterface;
 
 return static function (ThemeJson $themeJson, Presets $presets, ContainerInterface $container): void {
     // ...
@@ -444,8 +442,7 @@ declare(strict_types=1);
 
 namespace YourVendor\YourProject;
 
-use ItalyStrap\ThemeJsonGenerator\Application\Config\ThemeJson;
-use Psr\Container\ContainerInterface;
+use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\ThemeJson;use Psr\Container\ContainerInterface;
 
 return static function (ThemeJson $themeJson, ContainerInterface $container): void {
 // Utilize the $themeJson and $container for your configuration
@@ -465,8 +462,7 @@ declare(strict_types=1);
 
 namespace YourVendor\YourProject;
 
-use ItalyStrap\ThemeJsonGenerator\Application\Config\ThemeJson;
-use Psr\Container\ContainerInterface;
+use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\ThemeJson;use Psr\Container\ContainerInterface;
 
 return static function (ThemeJson $themeJson, ContainerInterface $container): void {
     /** @var SomeService $someService */
