+++ tests/_data/fixtures/basic-example.json.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace ItalyStrap\Tests;
 
-use ItalyStrap\ThemeJsonGenerator\Application\Config\ThemeJson;
-use ItalyStrap\ThemeJsonGenerator\Domain\Input\SectionNames;
+use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\SectionNames;
+use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\ThemeJson;
 
 return static function (ThemeJson $themeJson): void {
     $themeJson->merge([
